// apps/web/src/shared/api/generated/model/deliveryReadControllerHistoryListOrderState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type DeliveryReadControllerHistoryListOrderState =
  (typeof DeliveryReadControllerHistoryListOrderState)[keyof typeof DeliveryReadControllerHistoryListOrderState];

export const DeliveryReadControllerHistoryListOrderState = {
  active: "active",
  partially_shipped: "partially_shipped",
  shipped: "shipped",
  partially_received: "partially_received",
  received: "received",
  cancelled: "cancelled",
} as const;